requirements.txt
google-genai
python-dotenv
pypdf
fpdf2
chromadb
